python-server/requirements.txt
python-dotenv==1.0.1
vnstock3
pymongo==4.6.0
pandas==2.2.2
numpy==1.26.4
APScheduler==3.10.4
